% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonMar.R
\name{Lasso}
\alias{Lasso}
\title{Logistic probability model via penalized maximum likelihood}
\usage{
Lasso(xvec,y,xnew,lambda)
}
\arguments{
\item{xvec}{An input matrix. Each row is a vectorized predictor.}

\item{y}{Binary response variable.}

\item{xnew}{New predictors in the test data. Organized as a matrix with each row being a data point.}

\item{lambda}{The regularization penalty.}
}
\value{
The returned object is a list of components.

\code{B_est} - The estimated coefficient vector of linear predictor.

\code{prob} - The predicted probabilities for the test data.
}
\description{
Fit a logistic probability model based on Lasso penalty
}
